Fixed dy direction
dy was previously positive, which meant that the order in which the program calculated the mandelbrot set in was from the negative y's to the positive y's. However, when we want to save it as an image we want to save it going from the positive y's to the negatives. So the sign of dy was flipped. This means also that the starting position of y needs to be flipped.

Also removed some unnecessary memory placement calculations in the mandelbrot_grid function: saving to the array only needed incrementing instead of calculating the exact position of the array.